Extract vhm files and match extents to bhm / maxar

+++ src/data/filter_files.py
@@ -27,6 +27,7 @@ import numpy as np
 import os
 from os.path import expanduser
 import shutil
+import sys
 
 ###     USER INPUTS     ###
 
@@ -39,6 +40,9 @@ chunk_size = 100  # size to break full dataset into
 # filter on resolution (pixel_resolution) or dimensions (pixel_dimensions)
 filter_criterion = "pixel_dimensions"  # dimensions always recommended
 
+# whether to make zipped folders
+zip_files = False
+
 ### END OF USER INPUTS ###
 
 # get the home directory
@@ -49,6 +53,7 @@ basedir = homedir + "/data/UNICEF_data"
 
 # csv files
 csv_bhm = basedir + "/summary_data/BHM_pm.csv"
+csv_vhm = basedir + "/summary_data/VHM_pm_res.csv"
 csv_maxar = basedir + "/summary_data/maxar_pm.csv"
 
 # make a directory for the new file pairs
@@ -56,16 +61,18 @@ datapath = basedir + "/tim_maxar_bhm_final_pairs"
 
 # read in the dataframes
 df_bhm = pd.read_csv(csv_bhm)
+df_vhm = pd.read_csv(csv_vhm)
 df_maxar = pd.read_csv(csv_maxar)
 
 # get a filecode column which can be used to match the two dfs
 df_bhm["file_code"] = df_bhm.file_name.str[:8].str.split("-").str.join("").astype(int)
+df_vhm["file_code"] = df_vhm.file_name.str[:8].str.split("-").str.join("").astype(int)
 df_maxar["file_code"] = (
     df_maxar.file_name.str[:8].str.split("_").str.join("").astype(int)
 )
 
 # create pixel error x,y columns (difference between expected and actual pixel size)
-dfs = [df_bhm, df_maxar]
+dfs = [df_bhm, df_vhm, df_maxar]
 for df in dfs:
     df["pixel_err_x"] = np.abs(((df.right - df.left) / df.pixel_horiz) - 0.5)
     df["pixel_err_y"] = np.abs(((df.top - df.bottom) / df.pixel_vert) - 0.5)
@@ -75,6 +82,9 @@ if filter_criterion == "pixel_dimensions":
     df_bhm = df_bhm.loc[
         (df_bhm["pixel_horiz"] > pixel_small) & (df_bhm["pixel_vert"] > pixel_small)
     ]
+    df_vhm = df_vhm.loc[
+        (df_vhm["pixel_horiz"] > pixel_small) & (df_vhm["pixel_vert"] > pixel_small)
+    ]
     df_maxar = df_maxar.loc[
         (df_maxar["pixel_horiz"] > pixel_small) & (df_maxar["pixel_vert"] > pixel_small)
     ]
@@ -83,24 +93,36 @@ else:
     df_bhm = df_bhm.loc[
         (df_bhm["pixel_err_x"] <= err_tol) & (df_bhm["pixel_err_y"] <= err_tol)
     ]
+    df_vhm = df_vhm.loc[
+        (df_vhm["pixel_err_x"] <= err_tol) & (df_vhm["pixel_err_y"] <= err_tol)
+    ]
     df_maxar = df_maxar.loc[
         (df_maxar["pixel_err_x"] <= err_tol) & (df_maxar["pixel_err_y"] <= err_tol)
     ]
 
-# reduce the maxar dataset so it fully interesects with the bhm set
+# reduce the maxar and vhm datasets so they fully interesect with the bhm set
 df_maxar = df_maxar.loc[df_maxar["file_code"].isin(df_bhm["file_code"].values)]
+df_vhm = df_vhm.loc[df_vhm["file_code"].isin(df_bhm["file_code"].values)]
 
 # sort by file code and drop indices so the rows in both files line up
 df_bhm = df_bhm.sort_values("file_code", ascending=True)
+df_vhm = df_vhm.sort_values("file_code", ascending=True)
 df_maxar = df_maxar.sort_values("file_code", ascending=True)
 df_maxar = df_maxar.reset_index(drop=True)
 df_bhm = df_bhm.reset_index(drop=True)
+df_vhm = df_vhm.reset_index(drop=True)
+
 
 # add hyphen back into file codes
 df_bhm.file_code = (
     df_bhm.file_code.astype("str").str[:4]
     + "-"
     + df_bhm.file_code.astype("str").str[4:]
+)
+df_vhm.file_code = (
+    df_vhm.file_code.astype("str").str[:4]
+    + "-"
+    + df_vhm.file_code.astype("str").str[4:]
 )
 df_maxar.file_code = (
     df_maxar.file_code.astype("str").str[:4]
@@ -115,16 +137,31 @@ num_chunks = len(df_maxar) // chunk_size + 1
 chunk_sizes = chunk_size + np.zeros((num_chunks), dtype=int)
 chunk_sizes[-1] = num_chunks * chunk_size - len(df_maxar)
 
-for j in range(num_chunks):
-    subpath = datapath + "/pairs" + "_" + str(j)
-    try:
-        os.makedirs(subpath)
-        os.makedirs(subpath + "/maxar")
-        os.makedirs(subpath + "/bhm")
-    except FileExistsError:
-        pass
 
-    print("Zipping chunk", j + 1, " / ", num_chunks)
+def insert_row(idx, df, df_insert):
+
+    return (
+        df.iloc[
+            :idx,
+        ]
+        .append(df_insert)
+        .append(
+            df.iloc[
+                idx:,
+            ]
+        )
+        .reset_index(drop=True)
+    )
+
+
+for j in range(4, num_chunks):
+    print("Filtering batch", j + 1, "/", num_chunks)
+    subpath = datapath + "/pairs" + "_" + str(j)
+    for path in [subpath, subpath + "/maxar", subpath + "/bhm", subpath + "/vhm"]:
+        try:
+            os.makedirs(path)
+        except FileExistsError:
+            pass
     for k in range(chunk_sizes[j]):
         i = k + j * chunk_sizes[j]
 
@@ -146,8 +183,31 @@ for j in range(num_chunks):
             + ".tif"
         )
 
-        shutil.copyfile(old_maxar_name, new_maxar_name)
-        shutil.copyfile(old_bhm_name, new_bhm_name)
+        if df_bhm.file_code[i] != df_vhm.file_code[i]:
+            df_vhm = insert_row(i, df_vhm, df_bhm.loc[i])
+
+        new_vhm_name = subpath + "/vhm/vhm-" + str(df_vhm.file_code[i]) + ".tif"
+        old_vhm_name = (
+            basedir
+            + "/height-model-copy/"
+            + df_vhm.file_code[i][:4]
+            + "-VHM/VHM-"
+            + df_vhm.file_name[i]
+            + ".tif"
+        )
+
+        if not os.path.isfile(new_maxar_name):
+            shutil.copyfile(old_maxar_name, new_maxar_name)
+        if not os.path.isfile(new_bhm_name):
+            shutil.copyfile(old_bhm_name, new_bhm_name)
+        try:
+            shutil.copyfile(old_vhm_name, new_vhm_name)
+        except FileNotFoundError:
+            print("Skipping, no VHM file")
+            continue
 
     # zip the files
-    shutil.make_archive(datapath + "/maxar_bhm_pairs_" + str(j), "zip", subpath)
+    if zip_files:
+        shutil.make_archive(
+            datapath + "/maxar_bhm_vhm_triplets_" + str(j), "zip", subpath
+        )
